# backend/docs/swagger-output.json
{
  "openapi": "2.0",
  "info": {
    "title": "GLY VTU API",
    "description": "Auto-generated Swagger specification for GLY VTU backend.",
    "version": "3.4.4"
  },
  "servers": [
    {
      "url": "https://localhost:3000/"
    }
  ],
...
  "paths": {
...
    "/transactions/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/transactions/{id}/receipt": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/transaction/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api-docs.json": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "user-agent",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "origin",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sec-fetch-site",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sec-fetch-mode",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-internal-request",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/account/closure/cancel": {
      "get": {
        "description": "",
        "produces": [
          "text/csv"
        ],
        "parameters": [
          {
            "name": "user-agent",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "description": "Use: `Bearer <token>`"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}